refactor(eckhart): separate theme to fw and bld
