+++ core/embed/rust/src/ui/layout_eckhart/component/button.rs
@@ -38,7 +38,10 @@ pub struct Button {
 
 impl Button {
     const LINE_SPACING: i16 = 7;
+    #[cfg(not(feature = "bootloader"))]
     const SUBTEXT_STYLE: TextStyle = theme::label_menu_item_subtitle();
+    #[cfg(feature = "bootloader")]
+    const SUBTEXT_STYLE: TextStyle = theme::bootloader::TEXT_NORMAL;
 
     pub const fn new(content: ButtonContent) -> Self {
         Self {
